requirements.txt
pytesseract>=0.3.10
pdf2image>=1.16.3
Pillow>=10.0.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
scikit-learn>=1.3.0
numpy>=1.24.0
python-magic>=0.4.27
